Last commit as of now. Finished showing speed and other data into the speedos but skipped calibration now as GUI isn't the most important part. Alternatively, provided with a console where you can input any command of OBD2 and it will be executed. Please note that any command will be executed only 40 times as I've imposed a soft limit on it because my android phone is very slow and it is the original bottleneck.
If we skip some of GUI updates, we can achieve upto 5 queries per second on my simulator but with GUI updates on, I managed to achieve 2-3 queries per second which is after a lot of memory and algorithmic optimisations.
Anyways, updating the speedometer twice every second should be acceptable to the user as well as in cars too, this is the rate I guess?
Now, we can achieve 5 queries per second in a console output which can never be seen in the app but in the console window of android studio as Bluetooth communication is in the background thread.